优化连接释放与包头缓冲区复用，完善测试类型一致性
优化 MySqlConnection.Dispose，避免终结器线程关闭托管资源，提升线程安全性。SqlClient 新增 4 字节帧头缓冲区，替代池租借，提升性能与资源复用。ReadPacketAsync 复用包头缓冲区。单元测试中 _seq 断言统一强转为 Byte，保证类型一致。移除 AuthenticationTests 对 CONNECT_WITH_DB 标志的断言。

+++ UnitTest/SqlClientTests.cs
@@ -37,7 +37,7 @@ public class SqlClientTests
         Assert.Equal(len, rs.Length);
         //Assert.Equal(len, pk.Length);
         Assert.Equal(buf, pk.ReadBytes());
-        Assert.Equal(seq + 1, (Byte)client.GetValue("_seq")!);
+        Assert.Equal((Byte)(seq + 1), (Byte)client.GetValue("_seq")!);
         Assert.Equal(ms.Position, ms.Length);
     }
 
@@ -64,7 +64,7 @@ public class SqlClientTests
 
         Assert.Equal(len, rs.Length);
         Assert.Equal(buf, reader.ReadBytes(len));
-        Assert.Equal(seq + 1, (Byte)client.GetValue("_seq")!);
+        Assert.Equal((Byte)(seq + 1), (Byte)client.GetValue("_seq")!);
         Assert.Equal(ms.Position, ms.Length);
     }
 
@@ -91,7 +91,7 @@ public class SqlClientTests
 
         Assert.Equal(len, rs.Length);
         Assert.Equal(buf, reader.ReadBytes(len));
-        Assert.Equal(seq + 1, (Byte)client.GetValue("_seq")!);
+        Assert.Equal((Byte)(seq + 1), (Byte)client.GetValue("_seq")!);
         Assert.Equal(ms.Position, ms.Length);
     }
 
@@ -160,7 +160,7 @@ public class SqlClientTests
 
         Assert.True(rs.IsEOF);
         //Assert.Null(pk);
-        Assert.Equal(seq + 1, (Byte)client.GetValue("_seq")!);
+        Assert.Equal((Byte)(seq + 1), (Byte)client.GetValue("_seq")!);
         Assert.Equal(ms.Position, ms.Length);
     }
 
@@ -188,7 +188,7 @@ public class SqlClientTests
         Assert.Equal(pk.Length, rs.Length);
         Assert.Equal(len, (Int32)(rs.ToUInt32(0) & 0xFF_FFFF));
         Assert.Equal(seq, rs[3]);
-        Assert.Equal(seq + 1, (Byte)client.GetValue("_seq")!);
+        Assert.Equal((Byte)(seq + 1), (Byte)client.GetValue("_seq")!);
         Assert.Equal(buf, rs.ReadBytes(4, -1));
     }
 
@@ -210,7 +210,7 @@ public class SqlClientTests
         Assert.Equal(4 + buf.Length, rs.Length);
         Assert.Equal(len, (Int32)(rs.ToUInt32(0) & 0xFF_FFFF));
         Assert.Equal(seq, rs[3]);
-        Assert.Equal(seq + 1, (Byte)client.GetValue("_seq")!);
+        Assert.Equal((Byte)(seq + 1), (Byte)client.GetValue("_seq")!);
         Assert.Equal(buf, rs.ReadBytes(4, -1));
     }
